code cleaning
git-svn-id: svn://svn.code.sf.net/p/jmt/code/trunk@441 99a8c20d-b106-0410-ac6d-a91ca9afac8a

+++ src/jmt/engine/math/Erlang.java
@@ -29,43 +29,43 @@ package jmt.engine.math;
  * @version 1.0
  */
 public class Erlang {
-    
-    /**
-     * Computes Erlang-B formula (probability of blocking) with a 
-     * recursive method to avoid numeric instability
-     * @param traffic The total traffic offered (in erlangs) 
-     * that is <em>arrival rate * service demand</em>. Must be greater than zero.
-     * @param servers number of servers. Must be greater than one.
-     * @return computed value
-     * @throws IllegalArgumentException if <code>traffic <= 0</code> or <code>servers <= 0</code>
-     */
-    public static double erlangB(double traffic, int servers) throws IllegalArgumentException{
-        if (traffic < 0) {
-            throw new IllegalArgumentException("Traffic parameter must be greater than zero");
-        }
-        if (servers < 0) {
-            throw new IllegalArgumentException("Number of servers must be greater than zero");
-        }
-        double a = 1.0, inv = 1.0;
-        for (int i=0; i<servers; i++) {
-            a *= (double)(servers - i) / traffic;
-            inv += a;
-        }
-        return 1.0 / inv;
-    }
-    
-    /**
-     * Computes Erlang-C formula (probability of queueing) 
-     * basing on Erlang-B evaluation
-     * @param traffic The total traffic offered (in erlangs) 
-     * that is <em>arrival rate * service demand</em>. Must be greater than zero.
-     * @param servers number of servers. Must be greater than one.
-     * @return computed value
-     * @throws IllegalArgumentException if traffic <= 0 or servers <= 0
-     */
-    public static double erlangC(double traffic, int servers) throws IllegalArgumentException {
-        double b = erlangB(traffic, servers);
-        double rho = traffic / servers;
-        return b / (1 - rho + rho * b);
-    }
+
+	/**
+	 * Computes Erlang-B formula (probability of blocking) with a 
+	 * recursive method to avoid numeric instability
+	 * @param traffic The total traffic offered (in erlangs) 
+	 * that is <em>arrival rate * service demand</em>. Must be greater than zero.
+	 * @param servers number of servers. Must be greater than one.
+	 * @return computed value
+	 * @throws IllegalArgumentException if <code>traffic <= 0</code> or <code>servers <= 0</code>
+	 */
+	public static double erlangB(double traffic, int servers) throws IllegalArgumentException {
+		if (traffic < 0) {
+			throw new IllegalArgumentException("Traffic parameter must be greater than zero");
+		}
+		if (servers < 0) {
+			throw new IllegalArgumentException("Number of servers must be greater than zero");
+		}
+		double a = 1.0, inv = 1.0;
+		for (int i = 0; i < servers; i++) {
+			a *= (servers - i) / traffic;
+			inv += a;
+		}
+		return 1.0 / inv;
+	}
+
+	/**
+	 * Computes Erlang-C formula (probability of queueing) 
+	 * basing on Erlang-B evaluation
+	 * @param traffic The total traffic offered (in erlangs) 
+	 * that is <em>arrival rate * service demand</em>. Must be greater than zero.
+	 * @param servers number of servers. Must be greater than one.
+	 * @return computed value
+	 * @throws IllegalArgumentException if traffic <= 0 or servers <= 0
+	 */
+	public static double erlangC(double traffic, int servers) throws IllegalArgumentException {
+		double b = erlangB(traffic, servers);
+		double rho = traffic / servers;
+		return b / (1 - rho + rho * b);
+	}
 }
